+++ src/configuration.ts
@@ -4,27 +4,29 @@ import { Configuration, JsonConfiguration, TerminalWindow } from "./model";
 import * as path from "path";
 
 export async function getConfiguration(): Promise<Configuration> {
-  const keepExistingTerminalsOpen:
-    | boolean
-    | undefined = vscode.workspace
-    .getConfiguration("restoreTerminals")
-    .get("keepExistingTerminalsOpen");
+  const keepExistingTerminalsOpen: boolean | undefined =
+    vscode.workspace
+      .getConfiguration("turboCharger")
+      .get("keepExistingTerminalsOpen") ??
+    vscode.workspace
+      .getConfiguration("restoreTerminals")
+      .get("keepExistingTerminalsOpen");
 
-  const artificialDelayMilliseconds:
-    | number
-    | undefined = vscode.workspace
-    .getConfiguration("restoreTerminals")
-    .get("artificialDelayMilliseconds");
+  const artificialDelayMilliseconds: number | undefined =
+    vscode.workspace
+      .getConfiguration("turboCharger")
+      .get("artificialDelayMilliseconds") ??
+    vscode.workspace
+      .getConfiguration("restoreTerminals")
+      .get("artificialDelayMilliseconds");
 
-  const terminalWindows:
-    | TerminalWindow[]
-    | undefined = vscode.workspace
-    .getConfiguration("restoreTerminals")
-    .get("terminals");
+  const terminalWindows: TerminalWindow[] | undefined =
+    vscode.workspace.getConfiguration("turboCharger").get("terminals") ??
+    vscode.workspace.getConfiguration("restoreTerminals").get("terminals");
 
-  const runOnStartup: boolean | undefined = vscode.workspace
-    .getConfiguration("restoreTerminals")
-    .get("runOnStartup");
+  const runOnStartup: boolean | undefined =
+    vscode.workspace.getConfiguration("turboCharger").get("runOnStartup") ??
+    vscode.workspace.getConfiguration("restoreTerminals").get("runOnStartup");
 
   const configFromFile = await getConfigurationFromJsonFile();
   return {
@@ -49,15 +51,23 @@ async function getConfigurationFromJsonFile(): Promise<
   let configData: JsonConfiguration | undefined;
   //find any workspace with the config
   for (const folder of workspaceFolders) {
-    try {
-      const configFilePath = vscode.Uri.file(
-        path.join(folder.uri.fsPath, ".vscode", "restore-terminals.json")
-      );
-      const fileData = await vscode.workspace.fs.readFile(configFilePath);
-      const fileDataString = new TextDecoder("utf-8").decode(fileData);
-      configData = JSON.parse(fileDataString);
-    } catch (error) {
-      console.log("No config in workspace", folder, error);
+    for (const configFileName of [
+      "turbo-charger.json",
+      "restore-terminals.json",
+    ]) {
+      try {
+        const configFilePath = vscode.Uri.file(
+          path.join(folder.uri.fsPath, ".vscode", configFileName),
+        );
+        const fileData = await vscode.workspace.fs.readFile(configFilePath);
+        const fileDataString = new TextDecoder("utf-8").decode(fileData);
+        configData = JSON.parse(fileDataString);
+        break;
+      } catch (error) {
+        if (configFileName === "restore-terminals.json") {
+          console.log("No config in workspace", folder, error);
+        }
+      }
     }
   }
   if (!configData) return undefined;